requirements.txt
# pip install -r requirements.txt

Deprecated==1.2.13
Flask==2.0.3
loguru==0.6.0
lxml==4.8.0
matplotlib==3.5.1
numpy==1.21.5
onnx==1.11.0
onnx-simplifier==0.3.6
onnxoptimizer==0.2.6
onnxruntime-gpu==1.10.0
opencv-python==4.5.5.62
pika==1.2.0
Pillow==9.0.1
protobuf==3.19.4
pycuda==2021.1
pytools==2022.1.1
PyYAML==6.0
requests==2.27.1
scipy==1.7.3
tensorrt @ file:///home/ubuntu/TensorRT-8.2.3.0/python/tensorrt-8.2.3.0-cp37-none-linux_x86_64.whl
torch==1.10.2+cu113
torchaudio==0.10.2+cu113
torchvision==0.11.3+cu113
tqdm==4.62.3
uff @ file:///home/ubuntu/TensorRT-8.2.3.0/uff/uff-0.6.9-py2.py3-none-any.whl
yacs==0.1.8
